requirements.txt
cryptography==41.0.0
art==5.9
random-password-generator==2.2.0
tabulate==0.9.0
pwinput==1.0.3